requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.3
cymem==2.0.11
fastapi==0.119.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.4
packaging==25.0
pdfminer.six==20250506
preshed==3.0.10
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-docx==1.2.0
python-multipart==0.0.20
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.4.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.48.0
thinc==8.3.6
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
weasel==0.4.1
wrapt==2.0.0
